Move yapapi/executor/events.py to yapapi/events.py
