+++ yapapi/executor/ctx.py
@@ -8,7 +8,7 @@ from functools import partial
 from pathlib import Path
 from typing import Callable, Iterable, Optional, Dict, List, Tuple, Union, Any, Awaitable
 
-from .events import DownloadStarted, DownloadFinished
+from yapapi.events import DownloadStarted, DownloadFinished
 from ..props import NodeInfo
 from ..storage import StorageProvider, Source, Destination, DOWNLOAD_BYTES_LIMIT_DEFAULT
