+++ yapapi/services.py
@@ -19,9 +19,8 @@ if sys.version_info >= (3, 8):
 else:
     from typing_extensions import Final
 
-from yapapi import rest
+from yapapi import rest, events
 from yapapi.engine import _Engine, Job
-from yapapi.executor import events
 from yapapi.executor.ctx import WorkContext
 from yapapi.payload import Payload
 from yapapi.props import NodeInfo
